Added RocketPickResultEvent
