+++ src/main/java/net/mrscauthd/beyond_earth/entities/RocketTier3Entity.java
@@ -11,7 +11,6 @@ import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.MenuProvider;
 import net.minecraft.world.entity.*;
-import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
@@ -26,6 +25,7 @@ import io.netty.buffer.Unpooled;
 import net.mrscauthd.beyond_earth.events.Methods;
 import net.mrscauthd.beyond_earth.fluids.FluidUtil2;
 import net.mrscauthd.beyond_earth.guis.screens.rocket.RocketGui;
+import net.mrscauthd.beyond_earth.items.IRocketItem;
 import net.mrscauthd.beyond_earth.registries.ItemsRegistry;
 import net.mrscauthd.beyond_earth.registries.ParticlesRegistry;
 import net.mrscauthd.beyond_earth.registries.TagsRegistry;
@@ -43,25 +43,8 @@ public class RocketTier3Entity extends IRocketEntity {
 	}
 
 	@Override
-	public ItemStack getPickResult() {
-		ItemStack itemStack = new ItemStack(ItemsRegistry.TIER_3_ROCKET_ITEM.get(), 1);
-		itemStack.getOrCreateTag().putInt(BeyondEarthMod.MODID + ":fuel", this.getEntityData().get(FUEL));
-		itemStack.getOrCreateTag().putInt(BeyondEarthMod.MODID + ":buckets", this.getEntityData().get(BUCKETS));
-
-		return itemStack;
-	}
-
-	@Override
-	protected void spawnRocketItem() {
-		if (!level.isClientSide) {
-			ItemStack itemStack = new ItemStack(ItemsRegistry.TIER_3_ROCKET_ITEM.get(), 1);
-			itemStack.getOrCreateTag().putInt(BeyondEarthMod.MODID + ":fuel", this.getEntityData().get(FUEL));
-			itemStack.getOrCreateTag().putInt(BeyondEarthMod.MODID + ":buckets", this.getEntityData().get(BUCKETS));
-
-			ItemEntity entityToSpawn = new ItemEntity(level, this.getX(), this.getY(), this.getZ(), itemStack);
-			entityToSpawn.setPickUpDelay(10);
-			level.addFreshEntity(entityToSpawn);
-		}
+	protected IRocketItem getRocketItem() {
+		return ItemsRegistry.TIER_3_ROCKET_ITEM.get();
 	}
 
 	@Override
